Login form and Registration form with redirect + routs added
