+++ client/src/components/RegistrationForm/index.js
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+import Redirect from "react-router-dom/es/Redirect";
 import { makeStyles } from "@material-ui/core";
 import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
@@ -7,8 +8,14 @@ import InputLabel from "@material-ui/core/InputLabel";
 import OutlinedInput from "@material-ui/core/OutlinedInput";
 import InputAdornment from "@material-ui/core/InputAdornment";
 import IconButton from "@material-ui/core/IconButton";
+import Fade from "@material-ui/core/Fade";
+import Modal from "@material-ui/core/Modal";
+import Backdrop from "@material-ui/core/Backdrop";
 import { Visibility, VisibilityOff } from "@material-ui/icons";
+import ModalHeader from "../common/Modal/ModalHeader";
 // import axios from 'axios';
+// import AccountCircleSharpIcon from '@material-ui/icons/AccountCircleSharp';
+
 import theme from "../../theme";
 
 const useStyles = makeStyles({
@@ -42,6 +49,21 @@ const useStyles = makeStyles({
       backgroundColor: theme.palette.primary.dark,
     },
   },
+  modal: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  paper: {
+    backgroundColor: theme.palette.background.paper,
+    border: `5px solid ${theme.palette.primary.main}`,
+    borderRadius: "8px",
+    boxShadow: theme.shadows[5],
+    minWidth: "40vw",
+  },
+  icon: {
+    colors: theme.palette.primary.main,
+  },
 });
 
 const RegistrationForm = () => {
@@ -56,6 +78,12 @@ const RegistrationForm = () => {
     address: "",
     isAdmin: false,
   });
+
+  const [open, setOpen] = useState(true);
+
+  const handleClose = () => {
+    setOpen(false);
+  };
 
   const [showPassword, setShowPassword] = useState(false);
 
@@ -73,6 +101,7 @@ const RegistrationForm = () => {
 
   const submitRegistration = e => {
     e.preventDefault();
+    handleClose();
     // const newCustomer = values;
     // console.log(newCustomer);
     // axios.post('http://localhost:5000/customers', newCustomer)
@@ -80,89 +109,119 @@ const RegistrationForm = () => {
   };
 
   return (
-    <div className={classes.wrapper}>
-      <h3 className={classes.title}>Registration Form</h3>
-      <form className={classes.flexContainer}>
-        <TextField
-          label="First Name"
-          variant="outlined"
-          value={values.firstName}
-          required
-          onChange={handleChange("firstName")}
-          className={classes.textField}
-        />
-        <TextField
-          label="Last Name"
-          variant="outlined"
-          value={values.lastName}
-          required
-          onChange={handleChange("lastName")}
-          className={classes.textField}
-        />
-        <TextField
-          label="Login"
-          variant="outlined"
-          value={values.login}
-          onChange={handleChange("login")}
-          required
-          className={classes.textField}
-        />
-        <TextField
-          label="Email"
-          variant="outlined"
-          value={values.email}
-          onChange={handleChange("email")}
-          required
-          className={classes.textField}
-        />
-
-        <FormControl className={classes.textField} variant="outlined" required>
-          <InputLabel>Password</InputLabel>
-          <OutlinedInput
-            type={showPassword ? "text" : "password"}
-            value={values.password}
-            onChange={handleChange("password")}
-            endAdornment={
-              <InputAdornment position="end">
-                <IconButton
-                  aria-label="toggle password visibility"
-                  onClick={handleClickShowPassword}
-                  onMouseDown={handleMouseDownPassword}
-                  edge="end"
-                >
-                  {values.showPassword ? <Visibility /> : <VisibilityOff />}
-                </IconButton>
-              </InputAdornment>
-            }
-            labelWidth={70}
-          />
-        </FormControl>
-
-        <TextField
-          label="Telephone"
-          variant="outlined"
-          type="number"
-          value={values.telephone}
-          onChange={handleChange("telephone")}
-          className={classes.textField}
-        />
-        <TextField
-          label="Address"
-          variant="outlined"
-          value={values.address}
-          onChange={handleChange("address")}
-          className={classes.textField}
-        />
-        <Button
-          text="Submit"
-          variant="outline"
-          onClick={submitRegistration}
-          className={classes.btn}
+    <>
+      {open ? (
+        <Modal
+          aria-labelledby="transition-modal-title"
+          aria-describedby="transition-modal-description"
+          className={classes.modal}
+          open={open}
+          onClose={handleClose}
+          closeAfterTransition
+          BackdropComponent={Backdrop}
+          BackdropProps={{
+            timeout: 500,
+          }}
         >
-          Registration
-        </Button>
-      </form>
-    </div>
+          <Fade in={open}>
+            <div className={classes.paper}>
+              <ModalHeader />
+              <div className={classes.wrapper}>
+                <h3 className={classes.title}>Registration Form</h3>
+                <form className={classes.flexContainer}>
+                  <TextField
+                    label="First Name"
+                    variant="outlined"
+                    value={values.firstName}
+                    required
+                    onChange={handleChange("firstName")}
+                    className={classes.textField}
+                  />
+                  <TextField
+                    label="Last Name"
+                    variant="outlined"
+                    value={values.lastName}
+                    required
+                    onChange={handleChange("lastName")}
+                    className={classes.textField}
+                  />
+                  <TextField
+                    label="Login"
+                    variant="outlined"
+                    value={values.login}
+                    onChange={handleChange("login")}
+                    required
+                    className={classes.textField}
+                  />
+                  <TextField
+                    label="Email"
+                    variant="outlined"
+                    value={values.email}
+                    onChange={handleChange("email")}
+                    required
+                    className={classes.textField}
+                  />
+                  <FormControl
+                    className={classes.textField}
+                    variant="outlined"
+                    required
+                  >
+                    <InputLabel>Password</InputLabel>
+                    <OutlinedInput
+                      type={showPassword ? "text" : "password"}
+                      value={values.password}
+                      onChange={handleChange("password")}
+                      endAdornment={
+                        <InputAdornment position="end">
+                          <IconButton
+                            aria-label="toggle password visibility"
+                            onClick={handleClickShowPassword}
+                            onMouseDown={handleMouseDownPassword}
+                            edge="end"
+                          >
+                            {values.showPassword ? (
+                              <Visibility />
+                            ) : (
+                              <VisibilityOff />
+                            )}
+                          </IconButton>
+                        </InputAdornment>
+                      }
+                      labelWidth={70}
+                    />
+                  </FormControl>
+                  <TextField
+                    label="Telephone"
+                    variant="outlined"
+                    type="number"
+                    value={values.telephone}
+                    onChange={handleChange("telephone")}
+                    className={classes.textField}
+                  />
+                  <TextField
+                    label="Address"
+                    variant="outlined"
+                    value={values.address}
+                    onChange={handleChange("address")}
+                    className={classes.textField}
+                  />
+                  <Button
+                    text="Submit"
+                    variant="outline"
+                    onClick={submitRegistration}
+                    className={classes.btn}
+                  >
+                    Registration
+                  </Button>
+                </form>
+              </div>
+            </div>
+          </Fade>
+        </Modal>
+      ) : (
+        <Redirect to="/" />
+      )}
+    </>
   );
 };
