+++ client/src/routes/Routes.js
@@ -5,6 +5,8 @@ import { connect } from "react-redux";
 import HomePage from "../pages/HomePage/HomePage";
 import Cart from "../pages/Cart/Cart";
 import Profiler from "../pages/Profiler/Profiler";
+import LoginForm from "../components/LoginForm";
+import RegistrationForm from "../components/RegistrationForm";
 
 const Routes = ({ isAuthenticated = false }) => {
   return isAuthenticated ? (
@@ -27,6 +29,12 @@ const Routes = ({ isAuthenticated = false }) => {
       </Route>
       <Route path="/cart">
         <Cart />
+      </Route>
+      <Route path="/login">
+        <LoginForm />
+      </Route>
+      <Route path="/registration">
+        <RegistrationForm />
       </Route>
       <Redirect to="/" />
     </Switch>
